+++ changelog.md
@@ -2,9 +2,12 @@
 
 ## WIP
 
+## [1.6.0] - 2022-01-22
+
+- Add instances of `FunctorWithIndex`, `FoldableWithIndex` and
+  `TraversableWithIndex` for `Vector n a`.
 - Drop support for GHC older than 8.10
-- Safe construction of vectors from linked lists
-  - https://github.com/expipiplus1/vector-sized/pull/88
+- Safe construction of vectors from linked lists (https://github.com/expipiplus1/vector-sized/pull/88)
 
 Thanks to @sheaf and @kozross